% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseProtease.R
\name{browseProtease}
\alias{browseProtease}
\title{Browse Protease on MEROPS}
\usage{
browseProtease(p, keytype)
}
\arguments{
\item{p}{a single protease}

\item{keytype}{UniprotID (default) or MEROPS.}
}
\value{
utils::browseURL
}
\description{
Opens relevant MEROPS (https://www.ebi.ac.uk/merops/) page with information on specified protease.
}
\examples{
\dontrun{
if (interactive()) {

# The following function calls opens in browser

browseProtease("P07339", keytype = "UniprotID")
browseProtease("A01.009", keytype = "MEROPS")

}
}

}
