% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{cleavages}
\alias{cleavages}
\alias{cleavages-Cleavages}
\alias{cleavages,Cleavages-method}
\title{Access resulting object from `findProteases` function.}
\usage{
\S4method{cleavages}{Cleavages}(x)
}
\arguments{
\item{x}{A `data.table` object.}
}
\value{
A `data.table` object.
}
\description{
`cleavages` returns a `data.table` with the
corresponding details derived from MEROPS.
}
\examples{
protein <- c("P02671", "P02671", "P68871", "P01011")
peptide <- c("FEEVSGNVSPGTR", "FVSETESR", "LLVVYPW", "ITLLSAL")
res <- findProtease(protein = protein, peptide = peptide,
organism = "Homo sapiens")
cleavages(res)
}
