% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{proteases,Cleavages-method}
\alias{proteases,Cleavages-method}
\title{Access resulting object from \code{findProteases} function.}
\usage{
\S4method{proteases}{Cleavages}(x)
}
\arguments{
\item{object}{A \code{data.table} object.}
}
\value{
A \code{data.table} object.
}
\description{
\code{proteases}, \code{substrates} and \code{cleavages} returns a \code{data.table} with the
corresponding details derived from MEROPS.
}
