\name{proteasy-package}
\alias{proteasy-package}
\alias{proteasy}
\docType{package}
\title{
\packageTitle{proteasy}
}
\description{This package utilizes data derived from the MEROPS database (https://www.ebi.ac.uk/merops/) and facilitates retrieving proteolytic enzymes data by mapping peptide termini to known sites where a protease cleaves.}
\author{
\packageAuthor{proteasy}

Maintainer: \packageMaintainer{proteasy}
}
\keyword{ package }
