% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showProtease.R
\name{showProtease}
\alias{showProtease}
\title{Show Cleaving Data for a Peptidase or Inhibitor by Uniprot accession}
\usage{
showProtease(protein)
}
\arguments{
\item{protein}{a vector of corresponding UniProt Accession IDs.}

\item{summarize}{if false (default), provides a detailed table of all associated cleaving events, otherwise outputs a summarized table and only includes reviewed (Uniprot) entries.}
}
\value{
S4 object Cleavages
}
\description{
Given a vector of proteins, finds which substrates they cleave.
}
\examples{
protein <- c("P98073", "P00734")
showProtease(protein = protein)

}
