% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{cleavages}
\alias{cleavages}
\alias{cleavages-Cleavages}
\alias{cleavages,Cleavages-method}
\title{Access resulting object from `findProteases` function.}
\usage{
\S4method{cleavages}{Cleavages}(x)
}
\arguments{
\item{x}{A `data.table` object.}
}
\value{
A `data.table` object.
}
\description{
`cleavages` returns a `data.table` with the
corresponding details derived from MEROPS.
}
