% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProtease.R
\name{findProtease}
\alias{findProtease}
\title{Find Proteases}
\usage{
findProtease(peptide, protein)
}
\arguments{
\item{protein}{a vector of corresponding UniProt Accession IDs}

\item{peptide}{a vector of amino acid sequences}

\item{start_pos}{(optional) vector of N-terminus position in protein sequence}

\item{end_pos}{(optional) vector of C-terminus position in protein sequence}

\item{organism}{name of organism}
}
\value{
S4 object proteasy_res
}
\description{
Given a vector of peptides and proteins, finds proteases acting on cleavage sites
}
\examples{
protein <- c('P00450','P05387','P80297')
peptide <- c('AEEIIWNYAPSGIDIFTK','APGSAAPAAGSAPAAAEEK','CAGSCKCKECKCTSCKKSCCSCCPVGCAKCAQG')

}
