% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{substrates}
\alias{substrates}
\alias{substrates-Cleavages}
\alias{substrates,Cleavages-method}
\title{Access resulting object from `findProteases` function.}
\usage{
\S4method{substrates}{Cleavages}(x)
}
\arguments{
\item{x}{A `data.table` object.}
}
\value{
A `data.table` object.
}
\description{
`substrates` returns a `data.table` with the
corresponding details derived from MEROPS.
}
